+++ betterrandom/src/main/java/io/github/pr0methean/betterrandom/prng/BaseRandom.java
@@ -16,6 +16,7 @@ import io.github.pr0methean.betterrandom.util.Java8Constants;
 import java.io.IOException;
 import java.io.InvalidObjectException;
 import java.io.ObjectInputStream;
+import java.nio.ByteBuffer;
 import java.util.List;
 import java.util.Random;
 import java.util.concurrent.atomic.AtomicLong;
@@ -72,7 +73,11 @@ public abstract class BaseRandom extends Random
    * actual internal state of the PRNG is stored elsewhere (since otherwise getSeed() would require
    * a slow type conversion).
    */
-  protected byte[] seed;
+  protected volatile byte[] seed;
+  /**
+   * A {@link ByteBuffer} that wraps {@link #seed} if requested; null otherwise.
+   */
+  @Nullable protected transient volatile ByteBuffer seedBuffer;
   /**
    * Set by the constructor once either {@link Random#Random()} or {@link Random#Random(long)} has
    * returned. Intended for {@link #setSeed(long)}, which may have to ignore calls while this is
@@ -767,6 +772,7 @@ public abstract class BaseRandom extends Random
   protected void setSeedInternal(final byte[] seed) {
     if ((this.seed == null) || (this.seed.length != seed.length)) {
       this.seed = seed.clone();
+      initSeedBuffer();
     } else {
       System.arraycopy(seed, 0, this.seed, 0, seed.length);
     }
@@ -780,11 +786,11 @@ public abstract class BaseRandom extends Random
    * @param seedLength the length of the new seed in bytes
    */
   protected void creditEntropyForNewSeed(int seedLength) {
+    final long effectiveBits = Math.min(seedLength, getNewSeedLength()) * 8L;
     long oldCount;
     do {
       oldCount = entropyBits.get();
-    } while (!entropyBits.compareAndSet(oldCount,
-        Math.max(oldCount, Math.min(seedLength, getNewSeedLength()) * 8L)));
+    } while (!entropyBits.compareAndSet(oldCount, Math.max(oldCount, effectiveBits)));
   }
 
   /**
@@ -794,13 +800,29 @@ public abstract class BaseRandom extends Random
     superConstructorFinished = true;
   }
 
+  /**
+   * If overridden to return true, {@link #seedBuffer} is initialized with a buffer that wraps
+   * {@link #seed}; otherwise, that field remains null.
+   * @return true to wrap seed in a ByteBuffer; false otherwise
+   */
+  protected boolean usesByteBuffer() {
+    return false;
+  }
+
   private void readObject(final ObjectInputStream in) throws IOException, ClassNotFoundException {
     in.defaultReadObject();
     initTransientFields();
+    initSeedBuffer();
     setSeedInternal(seed);
     SeedGenerator currentSeedGenerator = getSeedGenerator();
     if (currentSeedGenerator != null) {
       RandomSeederThread.add(currentSeedGenerator, this);
+    }
+  }
+
+  private void initSeedBuffer() {
+    if (usesByteBuffer()) {
+      seedBuffer = ByteBuffer.wrap(this.seed);
     }
   }